theme & auto send options

+++ client/src/App.tsx
@@ -7,6 +7,9 @@ import ChatInput from './components/Chat/ChatInput';
 import ChatHead from './components/Chat/ChatHead';
 import NavPanel from './components/NavPanel';
 import AuthPanel from './components/NavPanel/AuthPanel';
+import OptionsPanel from './components/NavPanel/OptionsPanel';
+import ThemeToggle from './components/NavPanel/ThemeToggle';
+import AutoSendToggle from './components/NavPanel/AutoSendToggle';
 import ChatPanel from './components/NavPanel/ChatPanel';
 import Toast from './components/Toast';
 
@@ -32,23 +35,53 @@ interface Toast {
 function App() {
   const [chats, setChats] = useState<ChatModel[]>([]);
   const [currentChatId, setCurrentChatId] = useState<string | null>(null);
-  const currentChatIdRef = useRef<string | null>(null);
+  const currentChatIdRef = useRef<string | null>(currentChatId);
   currentChatIdRef.current = currentChatId;
 
   const [user, setUser] = useState<User | null>(null);
   const [isChecking, setIsChecking] = useState(true);
 
   const [searchTerm, setSearchTerm] = useState('');
-
   const [toasts, setToasts] = useState<Toast[]>([]);
+  const [botTypingChatId, setBotTypingChatId] = useState<string | null>(null);
+
+  const [isAutoSendOn, setIsAutoSendOn] = useState(false);
+  const isAutoSendOnRef = useRef(isAutoSendOn);
+  isAutoSendOnRef.current = isAutoSendOn;
 
   const currentChat = useMemo(() => {
     return chats.find((c) => c._id === currentChatId);
   }, [chats, currentChatId]);
 
-  const filteredChats = chats.filter((chat) =>
-    chat.botName.toLocaleLowerCase().includes(searchTerm.toLocaleLowerCase()),
+  const filteredChats = useMemo(
+    () =>
+      chats.filter((chat) =>
+        chat.botName
+          .toLocaleLowerCase()
+          .includes(searchTerm.toLocaleLowerCase()),
+      ),
+    [chats, searchTerm],
   );
+
+  const getRandomChat = useCallback(() => {
+    if (chats.length <= 0) return null;
+    const randomIdx = (Math.random() * chats.length) | 0;
+    return chats[randomIdx];
+  }, [chats]);
+
+  const checkAuthStatus = async () => {
+    try {
+      const response = await fetch(`${url}/auth/status`, {
+        credentials: 'include',
+      });
+      const data = await response.json();
+      setUser(data.user);
+    } catch (err) {
+      console.error('Error checking auth status:', err);
+    } finally {
+      setIsChecking(false);
+    }
+  };
 
   useEffect(() => {
     checkAuthStatus();
@@ -66,8 +99,8 @@ function App() {
           },
           body: JSON.stringify({ chats: chatsToSave }),
         });
-      } catch (error) {
-        console.error('Error saving chats:', error);
+      } catch (err) {
+        console.error('Error saving chats:', err);
       }
     },
     [user],
@@ -92,10 +125,10 @@ function App() {
         setChats(fixedChats);
       } else {
         setChats(initChats);
-        await remoteSaveChats(initChats);
+        remoteSaveChats(initChats);
       }
-    } catch (error) {
-      console.error('Error fetching chats:', error);
+    } catch (err) {
+      console.error('Error fetching chats:', err);
     }
   }, [remoteSaveChats]);
 
@@ -107,32 +140,84 @@ function App() {
     }
   }, [user, fetchChats]);
 
-  const checkAuthStatus = async () => {
-    try {
-      const response = await fetch(`${url}/auth/status`, {
-        credentials: 'include',
+  const appendMessage = useCallback(
+    (message: MessageModel, chatId: string | null = currentChatId) => {
+      if (!chatId) return;
+      setChats((prev) => {
+        const newChats = prev.map((chat) =>
+          chat._id === chatId
+            ? { ...chat, messages: [...chat.messages, message] }
+            : chat,
+        );
+        remoteSaveChats(newChats);
+        return newChats;
       });
-      const data = await response.json();
-      setUser(data.user);
-    } catch (error) {
-      console.error('Error checking auth status:', error);
-    } finally {
-      setIsChecking(false);
-    }
-  };
+    },
+    [currentChatId, remoteSaveChats],
+  );
 
-  const appendMessage = (message: MessageModel) => {
-    if (!currentChatId) return;
-    setChats((prev) => {
-      const newChats = prev.map((chat) =>
-        chat._id === currentChatId
-          ? { ...chat, messages: [...chat.messages, message] }
-          : chat,
-      );
-      remoteSaveChats(newChats);
-      return newChats;
-    });
-  };
+  const fetchQuote = useCallback(
+    async (chatId: string | null = currentChatId) => {
+      setBotTypingChatId(chatId);
+      const prevChatId = chatId;
+      try {
+        const res = await fetch(`${url}/quote/in3sec`);
+        const data: { quote: Quote } = await res.json();
+        const message: MessageModel = {
+          _id: generateObjectId(),
+          isBot: true,
+          content: data.quote.content,
+          timestamp: new Date(),
+        };
+        appendMessage(message, chatId);
+
+        // previous reference was closed upon closure
+        const newChatId = currentChatIdRef.current;
+
+        if (prevChatId !== newChatId) {
+          const chat = chats.find((chat) => chat._id === prevChatId);
+          chat && showToast(message.content, chat.botName);
+        }
+      } catch (err) {
+        console.error('Error fetching quote:', err);
+      } finally {
+        setBotTypingChatId(null);
+      }
+    },
+    [appendMessage, chats, currentChatId],
+  );
+
+  useEffect(() => {
+    let timeoutId: number;
+
+    const cleanup = () => {
+      if (timeoutId) {
+        clearTimeout(timeoutId);
+        timeoutId = 0;
+      }
+    };
+
+    const autoSend = () => {
+      if (!isAutoSendOnRef.current) {
+        cleanup();
+        return;
+      }
+
+      const randomChat = getRandomChat();
+      randomChat && fetchQuote(randomChat._id);
+
+      const randomDelay = ((Math.random() * 5000) | 0) + 5000;
+      timeoutId = setTimeout(autoSend, randomDelay);
+    };
+
+    if (isAutoSendOnRef.current) {
+      autoSend();
+    } else {
+      cleanup();
+    }
+
+    return cleanup;
+  }, [isAutoSendOn, getRandomChat, fetchQuote]);
 
   const showToast = (message: string, botName: string) => {
     const id = generateObjectId();
@@ -141,27 +226,6 @@ function App() {
 
   const removeToast = (id: string) => {
     setToasts((prevToasts) => prevToasts.filter((toast) => toast.id !== id));
-  };
-
-  const fetchQuote = async () => {
-    const prevChatId = currentChatId;
-    const res = await fetch(`${url}/quote/in3sec`);
-    const data: { quote: Quote } = await res.json();
-    const message: MessageModel = {
-      _id: generateObjectId(),
-      isBot: true,
-      content: data.quote.content,
-      timestamp: new Date(),
-    };
-    appendMessage(message);
-
-    // previous reference was closed upon closure
-    const newChatId = currentChatIdRef.current;
-
-    if (prevChatId !== newChatId) {
-      const chat = chats.find((chat) => chat._id === prevChatId);
-      chat && showToast(message.content, chat.botName);
-    }
   };
 
   const handleSend = (messageStr: string) => {
@@ -181,23 +245,21 @@ function App() {
   };
 
   const remoteDeleteChat = async (chatId: string) => {
+    if (!user) return;
     try {
       await fetch(`${url}/chats/${chatId}`, {
         method: 'DELETE',
         credentials: 'include',
-        headers: {
-          'Content-Type': 'application/json',
-        },
       });
-    } catch (error) {
-      console.error('Error deleting chat:', error);
+    } catch (err) {
+      console.error('Error deleting chat:', err);
     }
   };
 
   const handleDeleteChat = (chatId: string) => {
     setChats((prev) => {
       const newChats = prev.filter((chat) => chat._id !== chatId);
-      remoteDeleteChat(chatId);
+      user && remoteDeleteChat(chatId);
       return newChats;
     });
   };
@@ -212,7 +274,7 @@ function App() {
     };
     setChats((prev) => {
       const newChats = [...prev, newChat];
-      remoteSaveChats(newChats);
+      user && remoteSaveChats(newChats);
       return newChats;
     });
   };
@@ -223,7 +285,7 @@ function App() {
       const newChats = prev.map((chat) =>
         chat._id === chatId ? { ...chat, botName: newBotName } : chat,
       );
-      remoteSaveChats(newChats);
+      user && remoteSaveChats(newChats);
       return newChats;
     });
   };
@@ -247,7 +309,7 @@ function App() {
             }
           : chat,
       );
-      remoteSaveChats(newChats);
+      user && remoteSaveChats(newChats);
       return newChats;
     });
   };
@@ -262,8 +324,8 @@ function App() {
         credentials: 'include',
       });
       setUser(null);
-    } catch (error) {
-      console.error('Error logging out:', error);
+    } catch (err) {
+      console.error('Error logging out:', err);
     }
   };
 
@@ -278,6 +340,13 @@ function App() {
               onLogin={handleLogin}
               onLogout={handleLogout}
             />
+            <OptionsPanel>
+              <ThemeToggle />
+              <AutoSendToggle
+                isAutoSendOn={isAutoSendOn}
+                toggleIsAutoSendOn={() => setIsAutoSendOn((iaso) => !iaso)}
+              />
+            </OptionsPanel>
             <ChatPanel
               query={searchTerm}
               setQuery={(q: string) => setSearchTerm(q)}
@@ -298,6 +367,8 @@ function App() {
             <Chat>
               <ChatHead botName={currentChat.botName} />
               <MessageList
+                botName={currentChat.botName}
+                isBotTyping={botTypingChatId === currentChat._id}
                 messages={currentChat.messages}
                 onEdit={(messageId: string, newContent: string) =>
                   handleMessageEdit(currentChat._id, messageId, newContent)
@@ -310,7 +381,6 @@ function App() {
           )}
         </section>
       </main>
-
       <div className='toast-container'>
         {toasts.map((toast) => (
           <Toast
